requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alpha-vantage==2.3.1
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozendict==2.3.10
frozenlist==1.4.1
html5lib==1.1
idna==3.7
joblib==1.3.2
lxml==4.9.3
multidict==6.0.5
multitasking==0.0.11
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
peewee==3.17.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
ratelimiter==1.2.0.post0
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1
yarl==1.9.4
yfinance==0.2.33
